fix(client): address PR review feedback
