+++ packages/core/client/src/schema-component/antd/record-picker/__tests__/visibleCache.test.ts
@@ -11,7 +11,7 @@ import {
   getRecordPickerVisibleCacheKey,
   getRecordPickerVisibleFromCache,
   setRecordPickerVisibleToCache,
-} from '../InputRecordPicker';
+} from '../visibleCache';
 
 describe('record picker visible cache', () => {
   test('should prefer field path as cache key', () => {
